Code cleanup for traits and fix phala world mock module to include new traits from equip updates
