+++ traits/src/lib.rs
@@ -10,6 +10,7 @@ pub mod resource;
 pub mod theme;
 pub mod egg;
 pub mod preorders;
+pub mod status_type;
 
 pub use base::{Base, BaseInfo};
 pub use part::{EquippableList, FixedPart, PartType, SlotPart};
@@ -33,7 +34,6 @@ pub mod primitives {
 	pub type ZIndex = u32;
 	pub type SerialId = u32;
 	pub type EraId = u128;
-	pub type EggType = u8;
 	pub type RaceType = u8;
 	pub type CareerType = u8;
 }
